requirements.txt
qbittorrent_api==2022.10.39
schedule==1.1.0
vyper_config==1.1.1
